use depth with universal push too
